move subaru logic to cpp file
